# src/main/resources/static/v3/api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: MII Feasibility Backend REST API
  description: todo
  contact:
    email: noreply@todo.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.2
...
servers:
  - url: https://to.be.defined
    variables:
      basePath:
        default: /api/v3
...
paths:
...
  /terminology/entries/{nodeId}:
    get:
      tags:
        - terminology
      summary: Get the ui tree information for a node
      operationId: "getEntry"
      parameters:
        - name: "nodeId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminologyEntry"
        404:
          description: Node not found
          content: {}
  /terminology/entries:
    get:
      tags:
        - terminology
      summary: "Search for termcodes containing the search string"
      operationId: "search"
      parameters:
        - name: "query"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "categoryId"
          in: "query"
          required: false
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/TerminologyEntry"
  /terminology/{contextualizedTermcodeId}/ui_profile:
    get:
      tags:
        - terminology
      summary: "Get a UI profile for a given concept termcode and context termcode"
      operationId: "getUiProfile"
      parameters:
        - name: contextualizedTermcodeId
          in: path
          description: A hashed value (UUID v3), calculated from context and termcode with a predefined namespace
          required: true
          schema:
            type: "string"
          example: 43159902-7171-30cf-ab7e-ee36986c37c6
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "string"
              example:
                name: Prozedur
                time_restriction_allowed: true

        404:
          description: No matching ui profile found
          content: {}
...
  /terminology/criteria-set/intersect:
    post:
      tags:
        - terminology
      summary: "Submit a list of termcodes with their contexts and check if they are contained in a criteria set defined by the query parameter"
      operationId: "intersect"
      parameters:
        - name: criteriaSetUrl
          in: query
          description: The Canonical URL of the criteria set to check in
          required: true
          schema:
            type: "string"
          example: http://fhir.de/CriteriaSet/bfarm/icd-10-gm
        - name: contextTermCodeHashList
          in: body
          description: "A list of ContextTermCodeHashes"
          required: true
          schema:
            type: array
            items:
              type: string
            example: ["b8318334-7039-33aa-8ed0-85898fffa40b", "a57f98cf-e71d-3695-ae88-59d2c651bf0f", "430654d1-a7df-3c42-a33a-7210ef372bfd"]
      responses:
        200:
          description: Ok, return the filtered list of contextualized termcodes that are in the criteria set provided via query param
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example: ["b8318334-7039-33aa-8ed0-85898fffa40b", "a57f98cf-e71d-3695-ae88-59d2c651bf0f"]
...
components:
  schemas:
...
    TermCode:
      description: The termCode defines a concept based on a coding system (i.e. LOINC). The triplet of code, system and version identify the concept.
      type: object
      required:
        - code
        - system
        - display
      properties:
        code:
          type: string
          example: 119373006
        system:
          type: string
          example: http://snomed.info/sct
        version:
          type: string
          example: http://snomed.info/sct/900000000000207008/version/20210731
        display:
          type: string
          example: Amniotic fluid specimen (specimen)
...
    TerminologyEntry:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        context:
          $ref: "#/components/schemas/TermCode"
        termCodes:
          type: "array"
          items:
            $ref: "#/components/schemas/TermCode"
        termCode:
          $ref: "#/components/schemas/TermCode"
        children:
          type: "array"
          items:
            $ref: "#/components/schemas/TerminologyEntry"
        leaf:
          type: "boolean"
        selectable:
          type: "boolean"
        display:
          type: "string"
        root:
          type: "boolean"
...
  securitySchemes:
    feasibility_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://to.be.defined/auth
          scopes:
            user: Feasibility user role
            admin: Feasibility admin role